+++ src/document/Document.js
@@ -1,9 +1,18 @@
 const parse = require('./index');
 const toMarkdown = require('../output/markdown');
 const toHtml = require('../output/html');
+const toJSON = require('../output/json');
 const toLatex = require('../output/latex');
-const defaults = require('../lib/defaults');
-// const Image = require('../section/image/Image');
+const setDefaults = require('../lib/setDefaults');
+
+const defaults = {
+  infoboxes: true,
+  tables: true,
+  lists: true,
+  citations: true,
+  images: true,
+  sentences: true,
+};
 
 //
 const Document = function(wiki, options) {
@@ -12,18 +21,24 @@ const Document = function(wiki, options) {
 };
 
 const methods = {
+  title : function() {
+    if (this.options.title) {
+      return this.options.title;
+    }
+    let guess = null;
+    //guess the title of this page from first sentence bolding
+    let sen = this.sentences(0);
+    if (sen) {
+      guess = sen.bolds(0);
+    }
+    return guess;
+  },
   isRedirect : function() {
     return this.data.type === 'redirect';
   },
   isDisambiguation : function() {
     return this.data.type === 'disambiguation';
   },
-  // redirectTo : function() {
-  //   return p
-  // },
-  // disambigpages : function() {
-  //   return p
-  // },
   categories : function(n) {
     if (typeof n === 'number') {
       return this.data.categories[n];
@@ -37,7 +52,7 @@ const methods = {
     if (typeof n === 'string') {
       let str = n.toLowerCase().trim();
       return arr.find((s) => {
-        return s.title.toLowerCase() === str;
+        return s.title().toLowerCase() === str;
       });
     }
     if (typeof n === 'number') {
@@ -120,24 +135,24 @@ const methods = {
     return this.data.coordinates || [];
   },
   plaintext : function(options) {
-    options = Object.assign(defaults, options || {});
+    options = setDefaults(options, defaults);
     let arr = this.sections().map(sec => sec.plaintext(options));
     return arr.join('\n\n');
   },
   markdown : function(options) {
-    options = Object.assign(defaults, options || {});
+    options = setDefaults(options, defaults);
     return toMarkdown(this, options);
   },
   latex : function(options) {
-    options = Object.assign(defaults, options || {});
+    options = setDefaults(options, defaults);
     return toLatex(this, options);
   },
   html : function(options) {
-    options = Object.assign(defaults, options || {});
+    options = setDefaults(options, defaults);
     return toHtml(this, options);
   },
-  json : function() {
-    return this.data;
+  json : function(options) {
+    return toJSON(this, options);
   },
   debug: function() {
     console.log('\n');
@@ -146,7 +161,7 @@ const methods = {
       for(let i = 0; i < sec.depth; i += 1) {
         indent = ' -' + indent;
       }
-      console.log(indent + (sec.title || '(Intro)'));
+      console.log(indent + (sec.title() || '(Intro)'));
     });
   }
 };
@@ -172,5 +187,6 @@ Object.keys(methods).forEach((k) => {
 Document.prototype.toHTML = Document.prototype.html;
 Document.prototype.isDisambig = Document.prototype.isDisambiguation;
 Document.prototype.toJson = Document.prototype.json;
+Document.prototype.references = Document.prototype.citations;
 
 module.exports = Document;