+++ src/lib/defaults.js
@@ -1,8 +1,0 @@
-module.exports = {
-  infoboxes: true,
-  tables: true,
-  lists: true,
-  citations: true,
-  images: true,
-  sentences: true,
-};