+++ tests/stress.test.js
@@ -1,14 +1,6 @@
 'use strict';
 var test = require('tape');
-var fs = require('fs');
-var path = require('path');
-var wtf = require('./lib');
-
-//read cached file
-var feadFile = function(file) {
-  file = file.replace(/ /g, '-');
-  return fs.readFileSync(path.join(__dirname, 'cache', file + '.txt'), 'utf-8');
-};
+var readFile = require('./lib/_cachedPage');
 
 test('stress-test-en', t => {
   var arr = [
@@ -97,8 +89,7 @@ test('stress-test-en', t => {
     'Alexander-Y-Type': true
   };
   arr.forEach(title => {
-    var markup = feadFile(title);
-    var doc = wtf(markup);
+    var doc = readFile(title);
     //basic is-valid tests for the page parsing
     t.ok(true, title);
     t.equal(doc.isRedirect(), false, ' - - not-redirect');
@@ -106,8 +97,8 @@ test('stress-test-en', t => {
     t.ok(doc.categories().length > 0, ' - - cat-length');
     t.ok(doc.sections().length > 0, ' - - section-length');
     var intro = doc.sections(0);
-    t.ok(intro.title === '', ' - - intro-title-empty');
-    t.ok(intro.depth === 0, ' - - depth=0');
+    t.ok(intro.title() === '', ' - - intro-title-empty');
+    t.ok(intro.indentation() === 0, ' - - depth=0');
     t.ok(intro.sentences().length > 0, ' - - sentences-length');
     t.ok(intro.sentences(0).text().length > 0, ' - - intro-text');
     t.ok(intro.sentences(0).text().match(/[a-z]/), ' - - intro-has words');
@@ -116,13 +107,13 @@ test('stress-test-en', t => {
     } else {
       t.ok(doc.citations().length > 0, title + ' has a citation');
     }
-    var plain = wtf(markup).plaintext();
+    var plain = doc.plaintext();
     t.ok(plain.length > 40, ' - - plaintext-length');
 
-    var md = wtf(markup).markdown();
+    var md = doc.markdown();
     t.ok(md.length > 40, ' - - markdown-length');
 
-    var html = wtf(markup).html();
+    var html = doc.html();
     t.ok(html.length > 40, ' - - html-length');
     t.ok(html.match(/\</), ' - - html-has tag');
   });